+++ website/__init__.py
@@ -1,0 +1,5 @@
+from app import app
+from views import auth
+
+# Register the auth blueprint with the Flask application
+app.register_blueprint(auth)
