+++ website/static/myscripts.js
@@ -1,0 +1,19 @@
+document.addEventListener("DOMContentLoaded", function() {
+    const passwordField = document.getElementById("passwordField");
+    const showPasswordCheckbox = document.getElementById("showPasswordCheckbox");
+
+    showPasswordCheckbox.addEventListener("change", function() {
+        if (showPasswordCheckbox.checked) {
+            passwordField.type = "text";
+        } else {
+            passwordField.type = "password";
+        }
+    });
+});
+
+
+
+
+
+
+
