+++ website/templates/signup.html
@@ -5,17 +5,16 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Sign Up</title>
         <link rel="stylesheet" href="../static/style.css">
-        <script src="myscripts.js"></script>
+        <script src="../static/myscripts.js"></script>
     </head>
     <body>
         <nav>
             <ul>
-                <li><a href="home.html">Home</a></li>
+                <li><a href="/">Home</a></li>
                 <li><a href="#">About</a></li>
                 <li><a href="#">Make your appointment</a></li>
-                <li><a href="login.html">Login</a></li>
-                <li><a href="signup.html">Sign Up</a></li>
-                <li><a href="#" id="signOutLink">Sign Out</a></li>
+                <li><a href="/login">Login</a></li>
+                <li><a href="/signup">Sign Up</a></li>
             </ul>
         </nav>
